+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/GamePackageIntegrityInfo.cs
@@ -3,7 +3,7 @@
 
 using Snap.Hutao.Web.Hoyolab.HoyoPlay.Connect.ChannelSDK;
 
-namespace Snap.Hutao.Service.Game.Package;
+namespace Snap.Hutao.Service.Game.Package.Advanced;
 
 internal sealed class GamePackageIntegrityInfo
 {
