+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/IGameAssetsOperationService.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) DGP Studio. All rights reserved.
 // Licensed under the MIT license.
 
-namespace Snap.Hutao.Service.Game.Package;
+namespace Snap.Hutao.Service.Game.Package.Advanced;
 
 internal interface IGameAssetsOperationService
 {
